libctf: DW_AT_bit_offset may be signed
In certain cases that seem hard to figure out or justify, though a
packed bitfield structure between 33 and 40 bits in size seems to do it,
DW_AT_bit_offset maybe emitted by the compiler as a signed, negative,
value.

It is possible that this would also apply to DW_AT_data_bit_offset, but
I cannot reproduce that and haven't changed it.